a bunch of cleanup
