+++ snlscrape/conftest.py
@@ -1,3 +1,6 @@
+"""
+Module for declaring custom diagnostic messages for assertion failures in pytest tests.
+"""
 
 def pytest_assertrepr_compare(op, left, right):
   if isinstance(left, set) and isinstance(right, set) and op == '<=':
